Setup global app state and configuration
This is not IDEAL, but will suffice for now. A context manager variable
would be incredible amounts of overhead, and the usage is more or less
read only in all consumers. We could enforce this via a freeze() method
on the Configuration dict that returns some kind of Proxy type that is
immutable, but that is also a nightmare for later, if/when we run into
threading/async/concurrency issues.
For now, let's just assume we don't modify state outside of initialization.