Add stale data notification, refactor config
Add the ability for data that's serialized to be marked as stale after a
configurable interval (default is 24h). CLI will display this as dim
asterisks next to the fields that are outdated.
The Configuration class and all of its tests has also been moved to its own module under exsophere.config, which should have been the way from day 1. I don't know why it stuck inside Inventory for so long.