Add timeout and maxthread options
ssh connection timeout is now set to 10 seconds by default globally, can
be changed via 'global_timeout' in options. Additionally, this can be
overriden per host via the  'connect_timeout' key in hosts, from the
inventory.
'max_threads' has also been added, with a default of 15, and directly corresponds to the max_workers parameter of the ThreadPoolExecutor under the hood.