Refactor Update data to make current optional
Offloads the presentation of the lack of Current version to any
consumers of the Update dataclass. The Dataclass should just allow the
current_version field to be None instead of forcing implementations
where new packages can appear as part of updates to fill the field with
a string of their own choosing.
Update unit tests to reflect this, as well as Host command for formatting as "(new)" during display