Apt Provider: Handle new packages in parsing
Allows the Apt provider to correctly identify and inventory updates that
are new (no currently installed version), brought in as dependencies.
The FreeBSD pkg provider does this as well, so it is a decent precedent.
New packages pulled in as part of the dependency graph will have None in their current version field, which is handled by all display layers correctly.