cli host: Refine column display for updates
The Source column will no longer allow wrapping as this creates horrors
with the headers when the terminal is certain sizes.
I would like to present the Source column with elipses but I don't know how to make it happen with Rich tables, as they do this on their own and there's seemingly no control over it.