cli host: Put wrap constraint on cell
Previously was on column, but this caused weirdness. As a workaround, we
explicitly constraint the cell through a Text renderable, which causes
it to be properly ellipsised, and not wrapped when the line contains
spaces.
This is better than force squashing every column so it fits, and satisfies the requirement.