UI inventory: Add ability to redraw the table
Sometimes, the dashboard can update host state. Generally, those actions
trigger a refresh of the view, but this may not flow cross screen. I've
added a keybind to trigger a redraw manually for now.
Generally, I would prefer if this occured on its own but the only quick and dirty option I can think of is to just repopulate the table on screen resume, which can get expensive if it goes off every time you select a row to inspect it.