UI: Remove keys from DataTables and Widgets
The naming rules for host.name and the key name are too different, and
sanitizing them is too error prone and a one way process. Instead, we'll
just let Textual generate its own keys, and use it to select the column
by index given how static our data formats are. This has been done  for
Dashboard and Inventory, as well as the HostDetails panel.
Resolves issue: BadIdentifier for hosts with dot #21