Rework dashboard tiles to use labels
The dashboard HostWidget tiles now use labels for their contents,
inside a Vertical container. This allows for more flexible styling
and layout, without having to inject carriage returns like some sort
of caveperson.

The tiles are now a fixed 10 lines tall, with description and version
lines able to wrap. Online status is anchored to the bottom of the tile
for a consistent look within the grid.

Grid is now capped at 8 columns, which is still arbitrary but
feels reasonable in general.