Add kernel package support to DNF Provider
Kernels on RedHat likes are often - if not always - slotted packages.
This would qualify as a bug even though it is a dnf implementation
detail, since it massively breaks user expectations.

The dnf/yum provider now query the latest kernel available on
repositories and compare to installed kernels to find out if there is
an available update.

Additional changes:

- Better logging around slotted package handling
- Fix some docstrings that were out of date
- Remove extraneous info logging, as caller will log the message(s) at
  that loglevel.
- Add unit tests for new functionality
- Massively refactor fixtures to avoid fragile connection run output
  ordering