Add minor validations for host config
The name and ip field must be present, as per the documentation, but
this is made explicit as to give a better exception message. We now ban
the @ character from the ip field since the 'user@host' form is not
explicitly supported across the codebase, just the library. This can
(and will) result in unexpected behavior, so we now simply bail and
inform the user that they should use the the 'username' field to specify
this at all.