redhat provider: Prevent duplicates in updates
In some configurations (this varies from flavor to flavor), updates can
end up being obtained by any or all of security, kernel and general
updates. The interface does not have a convenient way of doing
exclusions in a way that retains coverage equally on all flavors, so we
simply check whether or not updates are already found before creating
the objects.
This is independent of version clobbering.