Providers: return current version for kernel
I changed my mind and instead of following dnf's inane rule about
slotted packages not being upgrades, we'll follow the principle of least
surprise and present the package as an upgrade, because the users
(myself included) will not care. Additionally, we already do this for
supplementary kernel packages such as kernel-core, kernel-modules,
kernel-tools etc, so consistency seems like a good idea.