FreeBSD Provider: Support repo tag in pkg output
Recent versions of pkg have started including the package repository
name in square brackets in the output that we parse. This unfortunately
resulted in parsing being entirely broken in this case. This patch
supports both formats, and adds unit test fixtures for both, in order to
maximize version support coverage.
Repository name will now be reported if available, if not, previous "Packages Mirror" value will be returned.