FreeBSD Provider: require sudo for reposync
While the actual upgrade process updates repositories, the read-only
check does not, and this depends on periodic jobs to refresh the
repositories. It makes sense to allow a manual refresh for this provider
that defaults to being skipped with the stock SudoPolicy in exosphere, as there's ultimately little reason not to.