Add Initial OpenBSD provider
Initial addition of the OpenBSD provider.
This is fairly basic, but also as complete as it needs to be given the
underlying simplicity of OpenBSD's package management system.

It is not exceedingly performant at the moment, given how pkg_add does
not provide an easier to parse output for checking for updates, but
it works, and is acceptable for the time being.