Docs: Detail syspatch usage for OpenBSD provider
Detail the usage of syspatch and include in required tools, it was
missing as an oversight. It's part of the base distribution but should
be installed nonetheless.

While at it, I also expanded the explanation for why syspatch is
required at all.

uname was removed from the list of required commands as it's not
actually used by the provider, but was at some point during development.

Also, minor formatting changes  across document for consistency.