Add version check command and functionality
The 'version' subcommand allows to display the current version and check
PyPI to determine whether or not we are using the latest version.

Version printing to console has been relocated to a function in utils
for consistency. The --version parameter to the root command remains for
backwards compatibility.

A new option, 'update_checks', has been added to the config to allow
disabling of PyPI checks if desired.

Unit tests covering functionality added.