Docs: Document new ^x behavior, update screenshot
The `^x` keybind now performs a sync and refresh operation in sequence.
The documentation has been updated to reflect this change as well as
updating the inventory screen screenshot to show the new keybind.

A new note box has also been added to warn users that syncing may take
a long time depending on inventory size and host specifications.