UI: Hide mode switch bindings for current screen
In order to conserve screen space and reduce clutter in the footer, we
now hide the key mode switch key binding for the currently active
screen.

This is done in an absolutely hacky manner, by just overriding the
global binding with a local one, and binding it to a no-op action
at the app level (app.none).

This is not ideal, but I genuinely do not understand how to work the
Textual check_action system in this context (app + modes) to make this
nicer and reactive.

This works, and while not great, it's not horrifying either.