Add host name tab completion to REPL
This required a significant refactor of the completion logic, for the
better I think.

I'm not overly fond of the static list of commands and options to
determine whether or not they accept host names, but extracting that
dynamically from Typer commands would be its own nightmare, which I
am absolutely not interested in undertaking at this time.