Do not use yield for pre-computed lists
Return the list directly instead. This is more honest, easier to read,
has explicit typing, and conserves the generators only where they are
actually needed.

"yield everywhere" was an artifact of previous iterations where we
lazily sent matches up the chain, but that is no longer the case
since we need to gather all of them in multiple spots.