Use global colors in inventory panels
Where possible, DataTables do not take that kind of dynamic styling
at all, just Text objects and Rich Rederables, so we stick to general
"green" and "red" colors in there.

This unfortunately affects the main inventory table view as well as the
updates list in the Host Details panel.

Everything else uses the theme colors, however.