Add action to test against lowest dependencies
This is to avoid further surprise regressions when development strays
too far from the minimum supported versions and introduces breaking
changes unknowingly.

The idea is that if this suite fails, something in the pyproject needs
to be adjusted, or the code itself needs to be fixed to work with the
lowest supported versions.