Add basic integration tests for TUI
The functional, integration tests for the TUI are done using the actual
Textual testing framework, which implies a dependency on pytest-asyncio.

This so far serves mostly as a smoke test to ensure the TUI starts up,
navigation works, and all modal Screens can be composed without
throwing an exception.

Coupled with the recently added CI action to run the test suite
against the minimum versions defined in pyproject.toml, this _would_
have caught the Textual version compatibility error. (I actually tested
this very scenario).

This should ensure we at least have a basic catch all failure mode
for TUI regressions in the future.