Remove context managers for connections
Remove the context manager usage with 'with' for Fabric connection
objects in setup module and Provider implementations.

This functionality is better left to calling code and left outside of
the realm of responsibilities of these modules.

The larger plan is to allow for batching and reuse of connections
through the Exosphere lifecycle by not delegating this and instead
explicitly controlling it through the operations flow if necessary.